add output table for 'buildtest build history' command. When user runs this command it will report a summary of all builds by reading all of the 'build.json' files in $BUILDTEST_ROOT/var/.history and print a subset of the information that is relevant
